minhand/ml-service/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pymongo>=4.6.1
pandas>=2.1.4
numpy>=1.26.2
scikit-learn>=1.3.2
xgboost>=2.0.2
scipy>=1.11.4
joblib>=1.3.2
python-dotenv>=1.0.0
requests>=2.31.0
